+++ src/components/Editor/Editor.less
@@ -8,6 +8,11 @@
     margin-bottom: 15px;
   }
 
+  .readAllRules {
+    margin: 15px 0;
+    display: block;
+  }
+
   &__category {
     .ant-radio-group {
       display: flex;
